Adding BasicSkill, updating PhysicalSkill
