+++ src/main/java/io/khasang/gahelp/entity/BasicSkill.java
@@ -1,0 +1,50 @@
+package io.khasang.gahelp.entity;
+
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.MappedSuperclass;
+
+@MappedSuperclass
+public abstract class BasicSkill {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private long Id;
+
+    private String name;
+    private String description;
+    private int powerOfSkill;
+
+    public long getId() {
+        return Id;
+    }
+
+    public void setId(long id) {
+        Id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public int getPowerOfSkill() {
+        return powerOfSkill;
+    }
+
+    public void setPowerOfSkill(int powerOfSkill) {
+        this.powerOfSkill = powerOfSkill;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+}
